feat: Add RegistrationMessageDialog, enable message sending for registrations, and extract ContactForm
